Move 'get_connection()' convenience APIs to subpackage __init__.
