+++ gcloud/storage/test___init__.py
@@ -153,3 +153,24 @@ class Test_set_defaults(unittest2.TestCase):
         self.assertEqual(SET_PROJECT_CALLED, [PROJECT])
         self.assertEqual(SET_CONNECTION_CALLED, [CONNECTION])
         self.assertEqual(SET_BUCKET_CALLED, [BUCKET])
+
+
+class Test_get_connection(unittest2.TestCase):
+
+    def _callFUT(self, *args, **kw):
+        from gcloud.storage import get_connection
+        return get_connection(*args, **kw)
+
+    def test_it(self):
+        from gcloud import credentials
+        from gcloud.storage import SCOPE
+        from gcloud.storage.connection import Connection
+        from gcloud.test_credentials import _Client
+        from gcloud._testing import _Monkey
+        client = _Client()
+        with _Monkey(credentials, client=client):
+            found = self._callFUT()
+        self.assertTrue(isinstance(found, Connection))
+        self.assertTrue(found._credentials is client._signed)
+        self.assertEqual(found._credentials._scopes, SCOPE)
+        self.assertTrue(client._get_app_default_called)
